refactor: split frontend/app.js into modular components
