+++ frontend/js/ui.js
@@ -1,244 +1,341 @@
-// state is imported read-only for rendering context (e.g. perPage, weights display).
-// ui.js never calls setState() — all writes go through the calling module.
-import { getStars } from './utils.js';
-import { state } from './state.js';
+import { state, els } from './state.js';
 
-// ── Toast Notifications ───────────────────────────────────────────────────────
-
-export function showToast(message, type = 'info', duration = 3500) {
-  const container = _getOrCreateToastContainer();
-  const toast = document.createElement('div');
-  toast.className = `toast toast--${type}`;
-  toast.setAttribute('role', 'alert');
-  toast.innerHTML = `
-    <span class="toast__icon">${{ success:'✓', error:'✕', info:'ℹ', warning:'⚠' }[type] ?? 'ℹ'}</span>
-    <span class="toast__message">${_esc(message)}</span>
-    <button class="toast__close" aria-label="Dismiss">×</button>
-  `;
-  toast.querySelector('.toast__close').addEventListener('click', () => _dismiss(toast));
-  container.appendChild(toast);
-  requestAnimationFrame(() => toast.classList.add('toast--visible'));
-  setTimeout(() => _dismiss(toast), duration);
+// ── Toast ────────────────────────────────────────────────────────────
+export function toast(message, type = 'info') {
+    const el = document.createElement('div');
+    el.className = `toast ${type}`;
+    el.textContent = message;
+    els.toastContainer.appendChild(el);
+    setTimeout(() => {
+        el.style.opacity = '0';
+        el.style.transform = 'translateX(100%)';
+        el.style.transition = '300ms ease';
+        setTimeout(() => el.remove(), 300);
+    }, 3500);
 }
 
-function _dismiss(toast) {
-  toast.classList.remove('toast--visible');
-  toast.addEventListener('transitionend', () => toast.remove(), { once: true });
+// ── HTML Escaping (centralised — used by ui.js and recommendations.js) ──
+export function esc(str) {
+    return String(str ?? '')
+        .replace(/&/g, '&amp;')
+        .replace(/</g, '&lt;')
+        .replace(/>/g, '&gt;')
+        .replace(/"/g, '&quot;');
 }
 
-function _getOrCreateToastContainer() {
-  let c = document.getElementById('toast-container');
-  if (!c) {
-    c = document.createElement('div');
-    c.id = 'toast-container';
-    document.body.appendChild(c);
-  }
-  return c;
+// ── Stars & Badges ───────────────────────────────────────────────────
+export function renderStars(rating) {
+    const full = Math.floor(rating);
+    const half = rating - full >= 0.5;
+    let html = '';
+    for (let i = 0; i < 5; i++) {
+        if (i < full) html += '<span class="star filled">★</span>';
+        else if (i === full && half) html += '<span class="star filled">★</span>';
+        else html += '<span class="star">★</span>';
+    }
+    return html;
 }
 
-// ── Modals ────────────────────────────────────────────────────────────────────
-
-export function showModal(id) {
-  const modal = document.getElementById(id);
-  if (!modal) return;
-  modal.setAttribute('aria-hidden', 'false');
-  modal.classList.add('modal--visible');
-  modal.querySelector('[autofocus], input, button')?.focus();
-  document.body.classList.add('modal-open');
+export function sentimentBadge(score) {
+    if (score > 0.05) return '<span class="product-card__sentiment sentiment-positive">Positive</span>';
+    if (score < -0.05) return '<span class="product-card__sentiment sentiment-negative">Negative</span>';
+    return '<span class="product-card__sentiment sentiment-neutral">Neutral</span>';
 }
 
-export function hideModal(id) {
-  const modal = document.getElementById(id);
-  if (!modal) return;
-  modal.setAttribute('aria-hidden', 'true');
-  modal.classList.remove('modal--visible');
-  document.body.classList.remove('modal-open');
+export function categoryIcon(cat) {
+    const c = (cat || '').toLowerCase();
+    if (c.includes('book') || c.includes('fiction') || c.includes('literature')) return '📚';
+    if (c.includes('tech') || c.includes('computer') || c.includes('electro')) return '💻';
+    if (c.includes('music') || c.includes('audio')) return '🎵';
+    if (c.includes('movie') || c.includes('film') || c.includes('video')) return '🎬';
+    if (c.includes('game') || c.includes('toy')) return '🎮';
+    if (c.includes('food') || c.includes('kitchen') || c.includes('cook')) return '🍳';
+    if (c.includes('sport') || c.includes('fitness')) return '⚽';
+    if (c.includes('health') || c.includes('beauty')) return '💊';
+    if (c.includes('cloth') || c.includes('fashion')) return '👕';
+    if (c.includes('home') || c.includes('garden')) return '🏡';
+    return '📦';
 }
 
-export function initModalDismiss() {
-  document.addEventListener('keydown', (e) => {
-    if (e.key === 'Escape')
-      document.querySelectorAll('.modal--visible').forEach(m => hideModal(m.id));
-  });
-  document.addEventListener('click', (e) => {
-    if (e.target.classList.contains('modal--visible')) hideModal(e.target.id);
-  });
-}
-
-// ── Loading States ────────────────────────────────────────────────────────────
-
-export function setLoadingState(area, active) {
-  document.querySelectorAll(`[data-loading-area="${area}"]`)
-    .forEach(el => el.classList.toggle('is-loading', active));
-  document.querySelectorAll(`[data-loading-btn="${area}"]`)
-    .forEach(btn => (btn.disabled = active));
-}
-
-// ── Product Card Rendering ────────────────────────────────────────────────────
-
-export function renderProductCards(items, opts = {}) {
-  const gridId = opts.context === 'recommendations' ? 'recommendations-grid' : 'products-grid';
-  const grid   = document.getElementById(gridId);
-  if (!grid) return;
-
-  if (!items.length) {
-    grid.innerHTML = `<div class="empty-state"><p>${
-      opts.context === 'search' && opts.query
-        ? `No results for "<strong>${_esc(opts.query)}</strong>"`
-        : 'No items to display.'
-    }</p></div>`;
-    return;
-  }
-
-  grid.innerHTML = items.map(item => _buildCard(item, opts.context)).join('');
-
-  grid.querySelectorAll('.product-card').forEach(card => {
-    card.addEventListener('click', () => {
-      // Lazy import breaks any circular dep at parse time
-      import('./recommendations.js').then(({ showRecommendations }) => {
-        showRecommendations(card.dataset.title);
-      });
-    });
-    card.addEventListener('keydown', e => {
-      if (e.key === 'Enter' || e.key === ' ') {
-        e.preventDefault(); // stop space from scrolling the page
-        card.click();
-      }
-    });
-  });
-}
-
-function _buildCard(item, context) {
-  const title    = _esc(item.title ?? item.product_name ?? 'Untitled');
-  const category = _esc(item.category ?? '');
-  const price    = item.price != null ? `$${parseFloat(item.price).toFixed(2)}` : '';
-  const rating   = item.rating != null ? parseFloat(item.rating).toFixed(1) : null;
-  const score    = item.hybrid_score ?? item.score ?? null;
-
-  return `
-    <article class="product-card" data-title="${title}"
-             role="button" tabindex="0" aria-label="View recommendations for ${title}">
-      <div class="card__body">
-        <p class="card__category">${category}</p>
-        <h3 class="card__title">${title}</h3>
-        ${rating ? `
-          <div class="card-rating">
-            ${getStars(item.rating || 0)}
-            <span class="review-count">(${item.review_count || 0} reviews)</span>
-          </div>
-          ` : ''}
-        <div class="card__footer">
-          ${price ? `<span class="card__price">${price}</span>` : ''}
-          ${score != null && context === 'recommendations'
-            ? `<span class="card__score">Match: ${(score*100).toFixed(0)}%</span>` : ''}
+// ── Skeleton Cards ───────────────────────────────────────────────────
+export function createSkeletonCard() {
+    return `
+        <div class="product-card skeleton-card">
+            <div class="skeleton skeleton-image"></div>
+            <div class="product-info">
+                <div class="skeleton skeleton-title"></div>
+                <div class="skeleton skeleton-text"></div>
+                <div class="skeleton skeleton-text short"></div>
+                <div class="skeleton-footer">
+                    <div class="skeleton skeleton-price"></div>
+                    <div class="skeleton skeleton-button"></div>
+                </div>
+            </div>
         </div>
-      </div>
-    </article>
-  `;
+    `;
 }
 
-// ── Pagination ────────────────────────────────────────────────────────────────
-
-export function renderPagination(current, total, perPage, onPageChange) {
-  const container = document.getElementById('pagination');
-  if (!container) return;
-  const pages = Math.ceil(total / perPage);
-  if (pages <= 1) { container.innerHTML = ''; return; }
-
-  const range = _pageRange(current, pages);
-  container.innerHTML = range.map(p =>
-    p === '…'
-      ? `<span class="page-ellipsis">…</span>`
-      : `<button class="page-btn ${p === current ? 'page-btn--active' : ''}"
-               data-page="${p}" aria-label="Page ${p}">${p}</button>`
-  ).join('');
-
-  container.querySelectorAll('.page-btn').forEach(btn =>
-    btn.addEventListener('click', () => onPageChange(+btn.dataset.page))
-  );
+export function showSkeletons(container, count = 8) {
+    container.innerHTML = Array(count).fill('').map(() => createSkeletonCard()).join('');
 }
 
-function _pageRange(c, t) {
-  if (t <= 7) return Array.from({ length: t }, (_, i) => i + 1);
-  if (c <= 4) return [1,2,3,4,5,'…',t];
-  if (c >= t-3) return [1,'…',t-4,t-3,t-2,t-1,t];
-  return [1,'…',c-1,c,c+1,'…',t];
+// ── Page Meta ────────────────────────────────────────────────────────
+export function setPageMeta(title, description) {
+    document.title = title
+        ? `${title} — HybridRec`
+        : 'HybridRec — Smart Recommendations';
+    const metaDesc = document.querySelector('meta[name="description"]');
+    if (metaDesc && description) metaDesc.setAttribute('content', description);
 }
 
-// ── Upload & Build ────────────────────────────────────────────────────────────
+// ── Filters ──────────────────────────────────────────────────────────
+export function applyFilters(products) {
+    return products.filter((p) => {
+        const matchesCategory =
+            !state.filters.category || p.category === state.filters.category;
 
-export function bindUploadHandler(onSuccess) {
-  const input     = document.getElementById('file-upload');
-  const uploadBtn = document.getElementById('upload-btn');
-  if (!input || !uploadBtn) return;
+        const matchesRating =
+            !state.filters.rating || (p.rating || 0) >= Number(state.filters.rating);
 
-  uploadBtn.addEventListener('click', () => input.click());
-  input.addEventListener('change', async (e) => {
-    const file = e.target.files[0];
-    if (!file) return;
-    if (!['.csv','.json'].some(ext => file.name.toLowerCase().endsWith(ext))) {
-      showToast('Only CSV and JSON files are supported.', 'error'); return;
-    }
-    setLoadingState('upload', true);
-    try {
-      const form = new FormData();
-      form.append('file', file);
-      const res  = await fetch('/api/upload', { method: 'POST', body: form });
-      if (!res.ok) throw new Error((await res.json().catch(()=>({}))).detail ?? `Error ${res.status}`);
-      const data = await res.json();
-      showToast(`Uploaded ${data.rows ?? ''} rows successfully.`, 'success');
-      onSuccess?.(data);
-    } catch (err) {
-      showToast(err.message, 'error');
-    } finally {
-      setLoadingState('upload', false);
-      input.value = '';
-    }
-  });
-}
+        let sentiment = 'neutral';
+        if ((p.avg_sentiment || 0) > 0.05) sentiment = 'positive';
+        else if ((p.avg_sentiment || 0) < -0.05) sentiment = 'negative';
 
-export function bindBuildModelsHandler(onSuccess) {
-  document.getElementById('build-btn')
-    ?.addEventListener('click', async () => {
-      setLoadingState('build', true);
-      showToast('Building models… this may take a moment.', 'info', 8000);
-      try {
-        const res = await fetch('/api/build', { method: 'POST' });
-        if (!res.ok) throw new Error((await res.json().catch(()=>({}))).detail ?? `Error ${res.status}`);
-        showToast('Models built! Start searching.', 'success');
-        onSuccess?.(await res.json());
-      } catch (err) {
-        showToast(err.message, 'error');
-      } finally {
-        setLoadingState('build', false);
-      }
+        const matchesSentiment =
+            !state.filters.sentiment || sentiment === state.filters.sentiment;
+
+        return matchesCategory && matchesRating && matchesSentiment;
     });
 }
 
-// ── Status Poller ─────────────────────────────────────────────────────────────
-
-export function startStatusPoller(intervalMs = 30_000) {
-  const poll = async () => {
-    try {
-      const res  = await fetch('/api/status');
-      if (!res.ok) return;
-      const data = await res.json();
-      const countEl  = document.getElementById('product-count');
-      const statusEl = document.getElementById('model-status');
-      if (countEl) countEl.textContent = `${(data.product_count ?? 0).toLocaleString()} products`;
-      if (statusEl) {
-        statusEl.textContent = data.models_built ? 'Models ready' : 'Models not built';
-        statusEl.className   = `status-badge status-badge--${data.models_built ? 'ok' : 'warn'}`;
-      }
-    } catch { /* silent */ }
-  };
-  poll();
-  return setInterval(poll, intervalMs);
+export function populateCategoryFilter(products) {
+    const categories = [...new Set(products.map((p) => p.category).filter(Boolean))];
+    els.categoryFilter.innerHTML = `
+        <option value="">All Categories</option>
+        ${categories.map((cat) => `<option value="${esc(cat)}">${esc(cat)}</option>`).join('')}
+    `;
 }
 
-// ── Escape helper ─────────────────────────────────────────────────────────────
+export function loadPreferences() {
+    const saved = localStorage.getItem('userPreferences');
+    if (!saved) return;
+    try {
+        const prefs = JSON.parse(saved);
+        state.filters.category = prefs.category || '';
+        state.filters.rating = prefs.rating || '';
+        state.filters.sentiment = prefs.sentiment || '';
+        els.categoryFilter.value = state.filters.category;
+        els.ratingFilter.value = state.filters.rating;
+        els.sentimentFilter.value = state.filters.sentiment;
+    } catch (err) {
+        console.warn('Failed to load preferences:', err);
+    }
+}
 
-export function escapeHtml(str) {
-  return String(str ?? '').replace(/&/g,'&amp;').replace(/</g,'&lt;')
-    .replace(/>/g,'&gt;').replace(/"/g,'&quot;');
+export function savePreferences() {
+    const prefs = {
+        category: state.filters.category,
+        rating: state.filters.rating,
+        sentiment: state.filters.sentiment,
+    };
+    localStorage.setItem('userPreferences', JSON.stringify(prefs));
+}
+
+// ── Wishlist ─────────────────────────────────────────────────────────
+export function getWishlist() {
+    return JSON.parse(localStorage.getItem('wishlist')) || [];
+}
+
+export function saveWishlist(items) {
+    localStorage.setItem('wishlist', JSON.stringify(items));
+}
+
+export function isWishlisted(title) {
+    return getWishlist().some((item) => item.title === title);
+}
+
+// ── Status ───────────────────────────────────────────────────────────
+export function updateStatus(cls, text) {
+    els.statusDot.className = `status-dot ${cls}`;
+    els.statusText.textContent = text;
+}
+
+// ── Heatmap ──────────────────────────────────────────────────────────
+export function renderHeatmap(labels, matrix) {
+    const n = labels.length;
+    const shortLabels = labels.map((l) => (l.length > 25 ? l.substring(0, 22) + '…' : l));
+
+    let html = `<div class="heatmap-grid" style="grid-template-columns: 140px repeat(${n}, 1fr); grid-template-rows: auto repeat(${n}, 1fr);">`;
+    html += '<div class="heatmap-cell heatmap-corner"></div>';
+
+    for (let j = 0; j < n; j++) {
+        html += `<div class="heatmap-cell heatmap-col-label" title="${esc(labels[j])}">${esc(shortLabels[j])}</div>`;
+    }
+
+    for (let i = 0; i < n; i++) {
+        html += `<div class="heatmap-cell heatmap-row-label" title="${esc(labels[i])}">${esc(shortLabels[i])}</div>`;
+        for (let j = 0; j < n; j++) {
+            const score = matrix[i][j];
+            const r = Math.round(255 - score * 200);
+            const g = Math.round(255 - score * 55);
+            const b = Math.round(255 - score * 200);
+            const bg = `rgb(${r}, ${g}, ${b})`;
+            const textColor = score > 0.6 ? '#fff' : 'var(--text)';
+            html += `<div class="heatmap-cell heatmap-value" style="background:${bg};color:${textColor};"
+                title="${esc(labels[i])} × ${esc(labels[j])}: ${score.toFixed(4)}">
+                ${score === 1 ? '1.0' : score.toFixed(2)}
+            </div>`;
+        }
+    }
+
+    html += '</div>';
+    els.heatmapContainer.innerHTML = html;
+}
+
+// ── Compare Bar (Side-by-Side) ───────────────────────────────────────
+export function updateCompareBar() {
+    let bar = document.getElementById('compare-bar');
+    if (!bar) {
+        bar = document.createElement('div');
+        bar.id = 'compare-bar';
+        bar.className = 'compare-bar';
+        document.body.appendChild(bar);
+    }
+    if (state.compareList.length === 0) {
+        bar.hidden = true;
+        return;
+    }
+    bar.hidden = false;
+    bar.innerHTML = `
+        <div class="compare-bar__items">
+            ${state.compareList.map((p) => `
+                <div class="compare-bar__item">
+                    <span>${esc(p.title.substring(0, 25))}${p.title.length > 25 ? '...' : ''}</span>
+                    <button onclick="window._removeFromCompare('${esc(p.title.replace(/'/g, "\\'"))}')">✕</button>
+                </div>
+            `).join('')}
+        </div>
+        <div class="compare-bar__actions">
+            <span class="compare-bar__count">${state.compareList.length}/3 selected</span>
+            <button class="compare-bar__btn" onclick="window._openComparePage()"
+                ${state.compareList.length < 2 ? 'disabled' : ''}>
+                Compare Now
+            </button>
+            <button class="compare-bar__clear" onclick="window._clearCompare()">Clear</button>
+        </div>
+    `;
+}
+
+export function openComparePage() {
+    if (state.compareList.length < 2) {
+        toast('Select at least 2 products to compare', 'info');
+        return;
+    }
+
+    let modal = document.getElementById('compare-modal');
+    if (!modal) {
+        modal = document.createElement('div');
+        modal.id = 'compare-modal';
+        modal.className = 'modal-overlay';
+        document.body.appendChild(modal);
+    }
+
+    const products = state.compareList;
+    modal.innerHTML = `
+        <div class="modal" style="max-width:900px;width:95%;">
+            <button class="modal__close" onclick="document.getElementById('compare-modal').hidden=true">&times;</button>
+            <h2 class="modal__title">Product Comparison</h2>
+            <div style="overflow-x:auto;margin-top:16px;">
+                <table class="compare-table">
+                    <thead>
+                        <tr>
+                            <th style="min-width:120px;">Attribute</th>
+                            ${products.map((p) => `<th style="min-width:180px;">${esc(p.title)}</th>`).join('')}
+                        </tr>
+                    </thead>
+                    <tbody>
+                        <tr>
+                            <td><strong>Category</strong></td>
+                            ${products.map((p) => `<td>${esc(p.category || 'N/A')}</td>`).join('')}
+                        </tr>
+                        <tr>
+                            <td><strong>Rating</strong></td>
+                            ${products.map((p) => `<td>⭐ ${(p.rating || 0).toFixed(1)}</td>`).join('')}
+                        </tr>
+                        <tr>
+                            <td><strong>Sentiment</strong></td>
+                            ${products.map((p) => {
+                                const s = p.avg_sentiment || 0;
+                                const label = s > 0.05 ? '😊 Positive' : s < -0.05 ? '😞 Negative' : '😐 Neutral';
+                                return `<td>${label}</td>`;
+                            }).join('')}
+                        </tr>
+                        <tr>
+                            <td><strong>Description</strong></td>
+                            ${products.map((p) => `<td style="font-size:12px;">${esc((p.description || 'N/A').substring(0, 100))}...</td>`).join('')}
+                        </tr>
+                    </tbody>
+                </table>
+            </div>
+        </div>
+    `;
+
+    modal.hidden = false;
+    modal.addEventListener('click', (e) => {
+        if (e.target === modal) modal.hidden = true;
+    });
+}
+
+// ── Back To Top ──────────────────────────────────────────────────────
+export function initBackToTop() {
+    const backToTop = document.getElementById('backToTop');
+    if (!backToTop) return;
+    backToTop.style.display = 'none';
+    window.addEventListener('scroll', () => {
+        backToTop.style.display = window.scrollY > 700 ? 'block' : 'none';
+    });
+    backToTop.addEventListener('click', () => {
+        window.scrollTo({ top: 0, behavior: 'smooth' });
+    });
+}
+
+// ── Theme Toggle ─────────────────────────────────────────────────────
+export function initTheme() {
+    const themeToggle = document.getElementById('theme-toggle');
+    if (!themeToggle) return;
+    const root = document.documentElement;
+    const savedTheme = localStorage.getItem('theme') || 'dark';
+    root.setAttribute('data-theme', savedTheme);
+    themeToggle.textContent = savedTheme === 'dark' ? '🌙' : '☀️';
+
+    themeToggle.addEventListener('click', () => {
+        const current = root.getAttribute('data-theme');
+        const next = current === 'dark' ? 'light' : 'dark';
+        root.setAttribute('data-theme', next);
+        localStorage.setItem('theme', next);
+        themeToggle.textContent = next === 'dark' ? '🌙' : '☀️';
+    });
+}
+
+// ── CSS Spin Animation ───────────────────────────────────────────────
+export function injectSpinStyle() {
+    const spinStyle = document.createElement('style');
+    spinStyle.textContent = `@keyframes spin { to { transform: rotate(360deg); } } .spin { animation: spin 1s linear infinite; }`;
+    document.head.appendChild(spinStyle);
+}
+
+// ── Debounce Helper ──────────────────────────────────────────────────
+export function debounce(func, delay) {
+    let timeout;
+    return function (...args) {
+        clearTimeout(timeout);
+        timeout = setTimeout(() => func.apply(this, args), delay);
+    };
+}
+
+// ── Upload file validation (case-insensitive) ────────────────────────
+export function isValidUploadFile(filename) {
+    // FIX: toLowerCase() ensures DATA.CSV and data.csv both pass
+    const lower = filename.toLowerCase();
+    return lower.endsWith('.csv') || lower.endsWith('.json');
 }
